Extract share content module Move SharedDataState and ShareDataViewModel to it
