+++ app/src/main/java/ir/mehdiyari/krypt/ui/KryptApp.kt
@@ -19,7 +19,7 @@ import ir.mehdiyari.krypt.R
 import ir.mehdiyari.krypt.core.designsystem.theme.KryptTheme
 import ir.mehdiyari.krypt.setting.ui.navigateToSettings
 import ir.mehdiyari.krypt.ui.data.navigateToData
-import ir.mehdiyari.krypt.ui.home.ShareDataViewModel
+import ir.mehdiyari.krypt.shareContent.ShareDataViewModel
 import ir.mehdiyari.krypt.features.auth.create_account.ROUTE_CREATE_ACCOUNT
 import ir.mehdiyari.krypt.features.auth.login.ROUTE_LOGIN
 import ir.mehdiyari.krypt.ui.media.MediaViewAction
@@ -34,7 +34,7 @@ fun KryptApp(
     hasAnyAccount: Boolean,
     onLockAppClicked: () -> Unit,
     onStopLocker: () -> Unit,
-    sharedDataViewModel: ShareDataViewModel,
+    sharedDataViewModel: ir.mehdiyari.krypt.shareContent.ShareDataViewModel,
 ) {
     KryptTheme {
         val snackbarHostState = remember { SnackbarHostState() }
