+++ app/src/main/java/ir/mehdiyari/krypt/ui/navigation/KryptNavHost.kt
@@ -8,7 +8,7 @@ import ir.mehdiyari.krypt.features.auth.login.navigateToLogin
 import ir.mehdiyari.krypt.setting.ui.settingsRoute
 import ir.mehdiyari.krypt.ui.KryptAppState
 import ir.mehdiyari.krypt.ui.data.dataScreen
-import ir.mehdiyari.krypt.ui.home.ShareDataViewModel
+import ir.mehdiyari.krypt.shareContent.ShareDataViewModel
 import ir.mehdiyari.krypt.ui.home.homeScreen
 import ir.mehdiyari.krypt.ui.home.navigateToHome
 import ir.mehdiyari.krypt.features.auth.create_account.createAccountScreen
@@ -30,7 +30,7 @@ fun KryptNaveHost(
     onShowSnackbar: suspend (String, String?) -> Boolean,
     startDestination: String,
     modifier: Modifier = Modifier,
-    sharedDataViewModel: ShareDataViewModel,
+    sharedDataViewModel: ir.mehdiyari.krypt.shareContent.ShareDataViewModel,
     onStopLocker: () -> Unit,
     onRestartApp: () -> Unit,
 ) {
