+++ settings.gradle.kts
@@ -33,3 +33,4 @@ include(":shared:files:files-logic")
 include(":features:settings")
 include(":shared:files:backup-logic")
 include(":features:auth:create-account")
+include(":shared:share-content")
