+++ shared/share-content/build.gradle.kts
@@ -1,0 +1,18 @@
+plugins {
+    id("krypt.android.library")
+    id("krypt.android.hilt")
+}
+
+android {
+    namespace = "ir.mehdiyari.krypt.shareContent"
+
+    defaultConfig {
+        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+    }
+}
+
+dependencies {
+
+    implementation(libs.androidx.lifecycle.viewmodelCompose)
+
+}
